move pre-installation package out of from license package
