+++ preinstallation/preinstallation.go
@@ -1,14 +1,14 @@
-// Package preinstall exposes an API to run necessary operations required in pre-install hook job of bootstrapping.
+// Package preinstallation exposes an API to run necessary operations required in pre-install hook job of bootstrapping.
 // While bootstrapping Tyk Stack, users need to provide a valida Tyk License key.
 // In the pre-hook installation, the helper functions defined in this package verifies the validity of the license.
-package preinstall
+package preinstallation
 
 import (
 	"errors"
 	"tyk/tyk/bootstrap/license"
 )
 
-// PreHookInstall runs all required License validity operations that are required in pre-install hook.
+// PreHookInstall runs all required license validation operations that are required in pre-install hook.
 func PreHookInstall() error {
 	dashboardLicenseKey, err := license.GetDashboardLicense()
 	if err != nil {
